requirements.txt
streamlit
python-dotenv
langchain
langchain-community
langchain-together
together
pypdf
PyPDF2
python-docx
google.genai
chromadb
sentence-transformers
